Test Scripts/Basic Posts/requirements.txt
requests
requests_oauthlib
atprototools
python-dotenv